Flushing cache
When using persistent object cache the drag/drop function doesn't work because the cache isn't flushed on save. It may save in the background, but the user won't see it until the cache is flushed.

This uses the stock WordPress functions to flush the cache making persistent object cache work.